Pickup and check_nearby_shelves functionality v1.01
However the shelves (x,y) in the check_nearby_shelves function has some problems. I have tried figuring it out, but idon'tknowifI'mblind. Every shelve has weirdly sat to x=0 and y=0.
The functions is just the start, since it doesnt require any seach input or from the list, but only searches if there are shelves y+1 / y-1 the robot. If there is, it tries to get that item and put it in the robot.
+ added functionality to the manuel_movement to type "t" for take. This runs first the check_nearby_shelves and then robot_item_pickup if a shelve was found.